splitted routes and controlers
